requirements.txt
# MCPDroid依赖需求
# 核心依赖
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
pillow>=9.0.0

# 高级功能依赖（可选）
# 图像识别依赖
opencv-python>=4.7.0
airtest>=1.2.7
# OCR依赖
pytesseract>=0.3.10
modelcontextprotocol==0.1.0